refactor(dartpad-preview): move source files from flame example into src folder #3719
