+++ pkgs/dartpad_preview/examples/flame_game/lib/src/components/brick.dart
