+++ pkgs/dartpad_preview/examples/flame_game/lib/src/components/paddle.dart
